MypageContent.vueをcomponentフォルダからviewフォルダへ移動
