+++ laugh_vue/laugh/src/router/index.ts
@@ -5,11 +5,7 @@ import DemoView from '../views/DemoView.vue'
 import MyPage from '../views/MyPage.vue'
 import LoginView from '../views/SingleLayout/Login.vue'
 import Main from '../views/Main.vue'
-import ProfileRegister_ from '../views/ProfileRegister_.vue'
-import ProfileRegister2 from '../views/ProfileRegister2.vue'
-import ProfileInit from '../views/ProfileInit.vue'
-// import ProfileEdit from '../views/ProfileEdit.vue'
-import MypageContent from '../components/MyPageContent.vue'
+import MypageContent from '../views/MypageContent.vue'
 import ProfileRegister from '../views/SingleLayout/ProfileRegister.vue'
 import OogiriDetail from '../views/OogiriDetailView.vue'
 import UserDetail from '../views/UserDetail.vue'
@@ -70,21 +66,6 @@ const router = createRouter({
           component: OogiriDetail,
         },
       ]
-    },
-    {
-      path: '/profile/register_/:address(.*)',
-      name: 'register',
-      component: ProfileRegister_
-    },
-    {
-      path: '/profile/register2/:address(.*)',
-      name: 'register2',
-      component: ProfileRegister2
-    },
-    {
-      path: '/profile/init',
-      name: 'registerInit',
-      component: ProfileInit
     },
   ]
 })
